split up utils, docstring improvements, and minor robustness tweaks to Mississippi scraper
