+++ workflow/python/covid19_scrapers/census/__init__.py
@@ -2,7 +2,7 @@ __all__ = ['CensusApi', 'FipsLookup', 'get_aa_pop_stats']
 
 from covid19_scrapers.census.census_api import CensusApi
 from covid19_scrapers.census.fips_lookup import FipsLookup
-from covid19_scrapers.utils import to_percentage
+from covid19_scrapers.utils.misc import to_percentage
 
 
 def get_aa_pop_stats(census_api, state, *, county=None, city=None):
