+++ workflow/python/covid19_scrapers/census/census_api.py
@@ -4,7 +4,7 @@ import logging
 import pandas as pd
 
 from covid19_scrapers.census.fips_lookup import FipsLookup
-from covid19_scrapers.utils import get_content_as_file, get_json
+from covid19_scrapers.utils.http import get_content_as_file, get_json
 
 
 _logger = logging.getLogger(__name__)
