+++ workflow/python/covid19_scrapers/census/fips_lookup.py
@@ -2,7 +2,7 @@ import logging
 
 import pandas as pd
 
-from covid19_scrapers.utils import get_content_as_file
+from covid19_scrapers.utils.http import get_content_as_file
 
 
 _logger = logging.getLogger(__name__)
