+++ workflow/python/covid19_scrapers/states/california.py
@@ -1,11 +1,11 @@
-from covid19_scrapers.utils import (
-    raw_string_to_int, table_to_dataframe, url_to_soup)
-from covid19_scrapers.scraper import ScraperBase
-
 import datetime
 import logging
 import re
 import string
+
+from covid19_scrapers.utils.html import table_to_dataframe, url_to_soup
+from covid19_scrapers.utils.parse import raw_string_to_int
+from covid19_scrapers.scraper import ScraperBase
 
 
 _logger = logging.getLogger(__name__)
