+++ workflow/python/covid19_scrapers/states/california_los_angeles.py
@@ -6,8 +6,10 @@ import re
 import pandas as pd
 
 from covid19_scrapers.census import get_aa_pop_stats
-from covid19_scrapers.utils import (
-    get_cached_url, raw_string_to_int, to_percentage, url_to_soup)
+from covid19_scrapers.utils.http import get_cached_url
+from covid19_scrapers.utils.parse import raw_string_to_int
+from covid19_scrapers.utils.misc import to_percentage
+from covid19_scrapers.utils.html import url_to_soup
 from covid19_scrapers.scraper import ScraperBase
