+++ workflow/python/covid19_scrapers/states/california_san_diego.py
@@ -7,8 +7,8 @@ from tabula import read_pdf
 
 from covid19_scrapers.census import get_aa_pop_stats
 from covid19_scrapers.scraper import ScraperBase
-from covid19_scrapers.utils import (
-    as_list, download_file, to_percentage)
+from covid19_scrapers.utils.http import download_file
+from covid19_scrapers.utils.misc import (as_list, to_percentage)
 
 
 _logger = logging.getLogger(__name__)
