+++ workflow/python/covid19_scrapers/states/colorado.py
@@ -6,7 +6,7 @@ from googleapiclient.discovery import build
 import pandas as pd
 
 from covid19_scrapers.scraper import ScraperBase
-from covid19_scrapers.utils import to_percentage
+from covid19_scrapers.utils.misc import to_percentage
 
 
 _logger = logging.getLogger(__name__)
