+++ workflow/python/covid19_scrapers/states/delaware.py
@@ -3,8 +3,10 @@ from datetime import datetime
 import re
 
 from covid19_scrapers.scraper import ScraperBase
-from covid19_scrapers.utils import (
-    raw_string_to_int, table_to_dataframe, to_percentage, url_to_soup)
+from covid19_scrapers.utils.html import table_to_dataframe, url_to_soup
+from covid19_scrapers.utils.misc import to_percentage
+from covid19_scrapers.utils.parse import raw_string_to_int
+
 
 _logger = logging.getLogger(__name__)
