+++ workflow/python/covid19_scrapers/states/idaho.py
@@ -5,7 +5,8 @@ from pytz import timezone
 from selenium.webdriver.common.by import By
 
 from covid19_scrapers.scraper import ScraperBase
-from covid19_scrapers.utils import get_content_as_file, to_percentage
+from covid19_scrapers.utils.http import get_content_as_file
+from covid19_scrapers.utils.misc import to_percentage
 from covid19_scrapers.webdriver import WebdriverRunner, WebdriverSteps
