+++ workflow/python/covid19_scrapers/states/maine.py
@@ -1,9 +1,10 @@
-from covid19_scrapers.utils import to_percentage, url_to_soup
-from covid19_scrapers.scraper import ScraperBase
-
 import logging
 import pandas as pd
 import re
+
+from covid19_scrapers.scraper import ScraperBase
+from covid19_scrapers.utils.html import url_to_soup
+from covid19_scrapers.utils.misc import to_percentage
 
 
 _logger = logging.getLogger(__name__)
