+++ workflow/python/covid19_scrapers/states/michigan.py
@@ -1,11 +1,13 @@
-from covid19_scrapers.utils import to_percentage, url_to_soup
-from covid19_scrapers.scraper import ScraperBase
-
 import datetime
 import logging
-import pandas as pd
 import re
 from urllib.parse import urljoin
+
+import pandas as pd
+
+from covid19_scrapers.scraper import ScraperBase
+from covid19_scrapers.utils.html import url_to_soup
+from covid19_scrapers.utils.misc import to_percentage
 
 
 _logger = logging.getLogger(__name__)
