+++ workflow/python/covid19_scrapers/states/montana.py
@@ -1,10 +1,12 @@
-from covid19_scrapers.utils import to_percentage, url_to_soup
-from covid19_scrapers.scraper import ScraperBase
-
 import datetime
 import logging
-import pytz
 import re
+
+import pytz
+
+from covid19_scrapers.scraper import ScraperBase
+from covid19_scrapers.utils.html import url_to_soup
+from covid19_scrapers.utils.misc import to_percentage
 
 
 _logger = logging.getLogger(__name__)
