+++ workflow/python/covid19_scrapers/states/nebraska.py
@@ -1,8 +1,9 @@
-from covid19_scrapers.scraper import ScraperBase
-from covid19_scrapers.utils import (
-    make_geoservice_stat, query_geoservice, to_percentage)
-
 import logging
+
+from covid19_scrapers.scraper import ScraperBase
+from covid19_scrapers.utils.arcgis import (
+    make_geoservice_stat, query_geoservice)
+from covid19_scrapers.utils.misc import to_percentage
 
 
 _logger = logging.getLogger(__name__)
