+++ workflow/python/covid19_scrapers/states/new_mexico.py
@@ -1,9 +1,10 @@
-from covid19_scrapers.utils import (get_content, get_json, to_percentage)
-from covid19_scrapers.scraper import ScraperBase
-
 import datetime
 import logging
 import re
+
+from covid19_scrapers.scraper import ScraperBase
+from covid19_scrapers.utils.http import get_content, get_json
+from covid19_scrapers.utils.misc import to_percentage
 
 
 _logger = logging.getLogger(__name__)
