+++ workflow/python/covid19_scrapers/states/new_york_city.py
@@ -8,7 +8,7 @@ import pandas as pd
 
 from covid19_scrapers.census import get_aa_pop_stats
 from covid19_scrapers.scraper import ScraperBase
-from covid19_scrapers.utils import to_percentage
+from covid19_scrapers.utils.misc import to_percentage
 
 
 _logger = logging.getLogger(__name__)
