+++ workflow/python/covid19_scrapers/states/texas_bexar.py
@@ -2,7 +2,8 @@ import logging
 
 from covid19_scrapers.census import get_aa_pop_stats
 from covid19_scrapers.scraper import ScraperBase
-from covid19_scrapers.utils import (query_geoservice, to_percentage)
+from covid19_scrapers.utils.arcgis import query_geoservice
+from covid19_scrapers.utils.misc import to_percentage
 
 
 _logger = logging.getLogger(__name__)
