+++ workflow/python/covid19_scrapers/states/virginia.py
@@ -1,8 +1,9 @@
-from covid19_scrapers.scraper import ScraperBase
-from covid19_scrapers.utils import to_percentage
-
 import logging
+
 import pandas as pd
+
+from covid19_scrapers.scraper import ScraperBase
+from covid19_scrapers.utils.misc import to_percentage
 
 
 _logger = logging.getLogger(__name__)
