+++ workflow/python/covid19_scrapers/states/wisconsin.py
@@ -1,7 +1,8 @@
-from covid19_scrapers.utils import (query_geoservice, to_percentage)
-from covid19_scrapers.scraper import ScraperBase
-
 import logging
+
+from covid19_scrapers.scraper import ScraperBase
+from covid19_scrapers.utils.arcgis import query_geoservice
+from covid19_scrapers.utils.misc import to_percentage
 
 
 _logger = logging.getLogger(__name__)
