+++ workflow/python/covid19_scrapers/utils/test_html.py
@@ -5,13 +5,8 @@ import pandas as pd
 import pytest
 import requests
 
-from covid19_scrapers.testing_utils import fake_webcache
+from covid19_scrapers.utils.testing import fake_webcache
 import covid19_scrapers.utils as utils
-
-
-def test_as_list():
-    assert utils.as_list('') == ['']
-    assert utils.as_list(['']) == ['']
 
 
 def test_url_to_soup():
@@ -19,32 +14,21 @@ def test_url_to_soup():
     with utils.UTILS_WEB_CACHE.with_instance(cache):
         # Good HTML
         session.add_response(content='<html><body>Body</body></html>')
-        soup = utils.url_to_soup('http://fake')
+        soup = utils.html.url_to_soup('http://fake')
         assert isinstance(soup, BeautifulSoup)
         assert soup.find('body') is not None
         assert soup.find('body').text == 'Body'
 
         # Bad HTML
         session.add_response(content='<html><>Body</body></html>')
-        soup = utils.url_to_soup('http://fake')
+        soup = utils.html.url_to_soup('http://fake')
         assert isinstance(soup, BeautifulSoup)
         assert soup.find('body') is None
 
         # Exception
         session.add_response(requests.RequestException('error'))
         with pytest.raises(requests.RequestException):
-            soup = utils.url_to_soup('http://fake')
-
-
-def test_raw_string_to_int():
-    assert 1 == utils.raw_string_to_int('1')
-    assert 1 == utils.raw_string_to_int('1a')
-    assert 1000 == utils.raw_string_to_int('1,000')
-    assert 100 == utils.raw_string_to_int('100%')
-    with pytest.raises(ValueError):
-        utils.raw_string_to_int('')
-    with pytest.raises(ValueError):
-        utils.raw_string_to_int('aaaa')
+            soup = utils.html.url_to_soup('http://fake')
 
 
 def test_find_all_links():
@@ -58,17 +42,17 @@ def test_find_all_links():
       <li><a href="link4">
   <body>
 </html>"""
-    with utils.UTILS_WEB_CACHE.with_instance(cache):
+    with utils.http.UTILS_WEB_CACHE.with_instance(cache):
         session.add_response(content=content)
-        links = utils.find_all_links('http://fake')
+        links = utils.html.find_all_links('http://fake')
         assert links == ['link1', 'link2', 'link3', 'link4', ]
 
         session.add_response(content=content)
-        links = utils.find_all_links('http://fake', '3')
+        links = utils.html.find_all_links('http://fake', '3')
         assert links == ['link3']
 
         session.add_response(content=content)
-        links = utils.find_all_links('http://fake', re.compile('[34]'))
+        links = utils.html.find_all_links('http://fake', re.compile('[34]'))
         assert links == ['link3', 'link4']
 
 
@@ -85,7 +69,8 @@ def test_table_to_dataframe():
     <tr><td>str3</td><td>100%</td><td> 100%</td></tr>
   </tbody>
 </table>"""
-    df = utils.table_to_dataframe(BeautifulSoup(table, features='lxml'))
+    df = utils.html.table_to_dataframe(BeautifulSoup(table,
+                                                     features='lxml'))
     print(df)
     expected = pd.DataFrame(
         columns=['str', 'int', 'flt'],
