+++ workflow/python/covid19_scrapers/utils/test_misc.py
@@ -1,0 +1,6 @@
+import covid19_scrapers.utils as utils
+
+
+def test_as_list():
+    assert utils.misc.as_list('') == ['']
+    assert utils.misc.as_list(['']) == ['']
